+++ bluefin/CHANGELOG.md
@@ -1,3 +1,5 @@
+* Add `Bluefin.CloneHandle` and `Bluefin.IO.withEffToIOCloneHandle`
+
 # 0.2.1.0
 
 * Add `Bluefin.DslBuilder`